requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
stripe==12.1.0
paypalrestsdk==1.13.3
gunicorn==23.0.0
email-validator==2.2.0
mailersend==0.5.0
Pyjwt==2.10.1
python-dotenv==1.1.0
python-jobspy==1.1.82
PyPDF2==3.0.1
docx==0.2.4
requests==2.31.0
fitz==0.0.1
PyMuPDF==1.26.3
psycopg2-binary==2.9.10
Flask-Migrate==4.1.0
Flask-Login==0.6.3
sentence-transformers==5.1.0
groq==0.29.0
fastapi==0.116.1
Werkzeug==3.1.3
SQLAlchemy==2.0.41
skillNer==1.0.3
pandas==2.3.0
spacy==3.8.7
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.8.0/en_core_web_md-3.8.0-py3-none-any.whl